Moved ElectroluxConfigEntry definition to coordinator.py
